+++ moo/mooutils/mooapp-ipc.c
@@ -227,7 +227,7 @@ moo_ipc_send (GObject    *sender,
               gssize      len)
 {
     GString *header;
-    guint id_len;
+    gsize id_len;
 
     g_return_if_fail (!sender || G_IS_OBJECT (sender));
     g_return_if_fail (id != NULL);
